Move PlanTraceObject to plan::global, move PlanProcessEdge/PlanScanObjects to scheduler::gc_work.
