+++ macros/trace_object/src/derive_impl.rs
@@ -57,7 +57,7 @@ pub(crate) fn generate_trace_object<'a>(
         fn trace_object<T: crate::plan::TransitiveClosure, const KIND: crate::policy::gc_work::TraceKind>(&self, __mmtk_trace: &mut T, __mmtk_objref: crate::util::ObjectReference, __mmtk_worker: &mut crate::scheduler::GCWorker<VM>) -> crate::util::ObjectReference {
             use crate::policy::space::Space;
             use crate::policy::gc_work::PolicyTraceObject;
-            use crate::plan::transitive_closure::PlanTraceObject;
+            use crate::plan::PlanTraceObject;
             #(#space_field_handler)*
             #parent_field_delegator
         }
@@ -119,7 +119,7 @@ pub(crate) fn generate_may_move_objects<'a>(
         #[inline(always)]
         fn may_move_objects<const KIND: crate::policy::gc_work::TraceKind>() -> bool {
             use crate::policy::gc_work::PolicyTraceObject;
-            use crate::plan::transitive_closure::PlanTraceObject;
+            use crate::plan::PlanTraceObject;
 
             false #(#space_handlers)* #parent_handler
         }
